+++ lib/src/button/ghost_button.dart
@@ -13,6 +13,7 @@ class GhostButton extends StatelessWidget {
     this.text,
     this.onPressed,
     this.borderRadius,
+    this.width,
   })  : assert(text == null || child == null, 'One of them must be null.'),
         assert(child != null || text != null, 'One of them must not be null.'),
         super(key: key);
@@ -30,13 +31,23 @@ class GhostButton extends StatelessWidget {
   /// Whether button has text content.
   final String? text;
 
+  /// If non-null, the corners of this box are rounded by this [BorderRadius].
+  ///
+  /// Applies only to boxes with rectangular shapes; ignored if [shape] is not
+  /// [BoxShape.rectangle].
+  ///
+  /// {@macro flutter.painting.BoxDecoration.clip}
   final BorderRadius? borderRadius;
+
+  /// if non-null, this box button will use specific width.
+  final double? width;
 
   @override
   Widget build(BuildContext context) {
     final enabled = onPressed != null;
     final theme = Theme.of(context);
     return Container(
+      width: width,
       decoration: BoxDecoration(
         color: Colors.transparent,
         borderRadius: borderRadius ?? BorderRadius.circular(12),