+++ lib/src/tag/_tag.dart
@@ -1,0 +1,8 @@
+// Created by Muhamad Fauzi Ridwan on 08/11/21.
+
+import 'package:flutter/material.dart';
+import 'package:ristek_material_component/ristek_material_component.dart';
+import 'package:ristek_material_component/src/spacer/_spacer.dart';
+
+part 'select_tag.dart';
+part 'tag.dart';